Converted the Button Tags R, P and S to the names ROCK, PAPER and SCISSORS for readability.
Simplified the UpdateTextandImage() Function by removing the switch statement and performing a simple resource lookup to set the image selected.

Added a nullable type for the playerChoice, set the Options Array and Random Funtion to readonly and added a check to verify the Button selected by the player contains a tag value.